tadded interface to the existing mapper class, renamed it to SqlMapper and created another mapper class names NoSqlMapper
